Created Statistics GUI

+++ src/main/scala/utility/Statistics.scala
@@ -14,22 +14,17 @@ object Statistics {
    * @param deathInBattle to sum
    * @param deathForNaturalCause to sum
    */
-  def update(foods: Int = 0,
-             attacks: Int = 0,
-             foodEaten: Int = 0,
-             deathInBattle: Int = 0,
-             deathForNaturalCause: Int = 0,
-            ): Unit = map =
+  def update(foods: Double = 0, attacks: Double = 0, foodEaten: Double = 0, deathInBattle: Double = 0, deathForNaturalCause: Double = 0): Unit = map =
     if (map isDefinedAt time)
       map + (time -> map(time).sum(StatisticsForTimeT(foods, attacks, foodEaten, deathInBattle, deathForNaturalCause)))
     else
       map + (time -> StatisticsForTimeT(foods, attacks, foodEaten, deathInBattle, deathForNaturalCause))
 
-  protected case class StatisticsForTimeT(foods: Int = 0,
-                                          attacks: Int = 0,
-                                          foodEaten: Int = 0,
-                                          deathInBattle: Int = 0,
-                                          deathForNaturalCause: Int = 0,
+  protected case class StatisticsForTimeT(foods: Double = 0,
+                                          attacks: Double = 0,
+                                          foodEaten: Double = 0,
+                                          deathInBattle: Double = 0,
+                                          deathForNaturalCause: Double = 0,
                                          ) {
     def sum(s: StatisticsForTimeT): StatisticsForTimeT =
       StatisticsForTimeT(foods + s.foods, attacks + s.attacks, foodEaten + s.foodEaten,
